+++ src/ezdxf/math/_vector.py
@@ -266,12 +266,14 @@ class Vec3:
                isclose(self._y, 0.) and \
                isclose(self._z, 0.)
 
-    def is_parallel(self, other: 'Vec3', *, abs_tol: float = 1e-12) -> bool:
+    def is_parallel(self, other: 'Vec3', *, rel_tol: float = 1e-9,
+                    abs_tol: float = 1e-12) -> bool:
         """ Returns ``True`` if `self` and `other` are parallel to vectors. """
         v1 = self.normalize()
         v2 = other.normalize()
-        return v1.isclose(v2, abs_tol=abs_tol) or v1.isclose(
-            -v2, abs_tol=abs_tol)
+        return v1.isclose(
+            v2, rel_tol=rel_tol, abs_tol=abs_tol) or v1.isclose(
+            -v2, rel_tol=rel_tol, abs_tol=abs_tol)
 
     @property
     def spatial_angle(self) -> float:
@@ -337,12 +339,15 @@ class Vec3:
         """ Returns ``True`` if vector is not ``(0, 0, 0)``. """
         return not self.is_null
 
-    def isclose(self, other: 'Vertex', *, abs_tol: float = 1e-12) -> bool:
-        """ Returns ``True`` if `self` is close to `other`. Uses :func:`math.isclose` to compare all axis. """
+    def isclose(self, other: 'Vertex', *, rel_tol: float = 1e-9,
+                abs_tol: float = 1e-12) -> bool:
+        """ Returns ``True`` if `self` is close to `other`.
+        Uses :func:`math.isclose` to compare all axis.
+        """
         x, y, z = self.decompose(other)
-        return math.isclose(self._x, x, abs_tol=abs_tol) and \
-               math.isclose(self._y, y, abs_tol=abs_tol) and \
-               math.isclose(self._z, z, abs_tol=abs_tol)
+        return math.isclose(self._x, x, rel_tol=rel_tol, abs_tol=abs_tol) and \
+               math.isclose(self._y, y, rel_tol=rel_tol, abs_tol=abs_tol) and \
+               math.isclose(self._z, z, rel_tol=rel_tol, abs_tol=abs_tol)
 
     def __eq__(self, other: 'Vertex') -> bool:
         """
@@ -716,11 +721,13 @@ class Vec2:
     def __bool__(self) -> bool:
         return not self.is_null
 
-    def isclose(self, other: 'VecXY', *, abs_tol: float = 1e-12) -> bool:
+    def isclose(self, other: 'VecXY', *, rel_tol: float = 1e-9,
+                abs_tol: float = 1e-12) -> bool:
         if not isinstance(other, Vec2):
             other = Vec2(other)
-        return math.isclose(self.x, other.x, abs_tol=abs_tol) and math.isclose(
-            self.y, other.y, abs_tol=abs_tol)
+        return math.isclose(
+            self.x, other.x, rel_tol=rel_tol, abs_tol=abs_tol) and math.isclose(
+            self.y, other.y, rel_tol=rel_tol, abs_tol=abs_tol)
 
     def __eq__(self, other: 'Vertex') -> bool:
         return self.__hash__() == hash(Vec2(other))